refactor: remove proxyImages function from content.get.js and move it to helper.js for better organization
