+++ server/api/entries/[entryId]/full-content.get.js
@@ -25,10 +25,14 @@ export default defineEventHandler(async (event) => {
   if (!entry) throw createError({ statusCode: 404, statusMessage: "Entry not found" });
   if (!entry.link) throw createError({ statusCode: 400, statusMessage: "Entry has no link" });
 
+  const feed = await repository.findFeedById(userId, entry.feedId);
+  if (!feed) throw createError({ statusCode: 404, statusMessage: "Feed not found" });
+
   const res = await downloadService.downloadText({ url: entry.link });
   if (!res) throw createError({ statusCode: 400, statusMessage: "Failed to download entry content" });
 
   const doc = new JSDOM(res.body, { url: entry.link });
   const reader = new Readability(doc.window.document);
-  return reader.parse();
+  const parsed = reader.parse();
+  return { content: proxyImages(parsed?.content || "", feed.htmlUrl) };
 });
